fix:✨ 팀 참가 유저 명단 조회 수정
