+++ src/main/java/com/goormdari/domain/user/presentation/UserController.java
@@ -51,24 +51,4 @@ public class UserController {
         return ResponseCustom.OK(userService.findCurrentStepById(userId));
     }
 
-    @Operation(summary = "팀 참가 유저 명단 조회", description = "같은 팀에 존재하는 유저 명단")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "조회 성공 ", content = {@Content(mediaType = "application/json", schema = @Schema(implementation =  findByTeamIdResponse.class))}),
-            @ApiResponse(responseCode = "400", description = "조회 실패", content = {@Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))}),
-    })
-    @GetMapping("/team-list")
-    public ResponseCustom<List<findByTeamIdResponse>> getTeamMember(
-            @Parameter(description = "Accesstoken을 입력해주세요.", required = true) @RequestHeader("Authorization") String token
-    ) {
-        if (token == null) {
-            throw new InvalidTokenException();
-        }
-
-        String jwt = token.startsWith("Bearer ") ? token.substring(7) : token;
-        if (!jwtUtil.validateToken(jwt)) {
-            throw new IllegalArgumentException("Invalid token");
-        }
-        Long userId = jwtUtil.extractId(jwt);
-        return ResponseCustom.OK(userService.findTeamByUserId(userId));
-    }
 }
